chore: improved autoloader, renamed Views
