+++ .gitignore
@@ -1,4 +1,3 @@
 /vendor/*
 composer.lock
 .env
-
