+++ app/View/components/user/planning/searchResults.php
