+++ app/View/components/user/planning/searchResultsUser.php
