+++ app/View/templates/user/planning.php
