+++ app/View/utils/global.php
